requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
certifi==2022.9.24
charset-normalizer==2.1.1
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorlog==6.7.0
datasets==2.7.1
dill==0.3.6
filelock==3.8.0
frozenlist==1.3.3
fsspec==2022.11.0
greenlet==2.0.1
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==4.13.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.14
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
optuna==3.0.3
packaging==21.3
pandas==1.5.2
pbr==5.11.0
pdoc3==0.10.0
Pillow==9.3.0
prettytable==3.5.0
pyarrow==10.0.1
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.44
stevedore==4.1.1
tokenizers==0.13.2
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
xxhash==3.1.0
yarl==1.8.1
zipp==3.11.0
